dependencies/imManager/requirements.txt
Pillow==5.4.1
requests==2.20.1